Delete and Update button upgrade

+++ app/recipes/[id]/page.jsx
@@ -1,170 +1,3 @@
-// "use client";
-// import { useEffect, useState } from "react";
-// import { doc, getDoc } from "firebase/firestore";
-// import { db } from "../../../firebase";
-// import { useRouter } from "next/navigation";
-
-// export default function RecipeDetails({ params }) {
-//   const [recipeDetails, setRecipeDetails] = useState(null);
-//   const recipeId = params.id;
-//   const router = useRouter();
-
-//   // useEffect(() => {
-//   //   const fetchRecipeDetails = async () => {
-//   //     const recipeDocRef = doc(db, "Recipes", recipeId);
-//   //     const recipeDoc = await getDoc(recipeDocRef);
-
-//   //     if (recipeDoc.exists()) {
-//   //       setRecipeDetails(recipeDoc.data());
-//   //     } else {
-//   //       console.log("Recipe not found");
-//   //     }
-//   //   };
-
-//   //   fetchRecipeDetails();
-//   // }, [recipeId]);
-
-//   useEffect(() => {
-//     const fetchRecipeDetails = async () => {
-//       try {
-//         const recipeDocRef = doc(db, "Recipes", recipeId);
-//         const recipeDoc = await getDoc(recipeDocRef);
-
-//         if (recipeDoc.exists()) {
-//           const recipeData = recipeDoc.data();
-
-//           // Retrieve user information based on userId
-//           const userDocRef = doc(db, "Users", recipeData.userId);
-//           const userDoc = await getDoc(userDocRef);
-
-//           if (userDoc.exists()) {
-//             const userData = userDoc.data();
-
-//             // Merge recipe and user data
-//             const mergedData = {
-//               ...recipeData,
-//               postedByUserName: userData.name,
-//             };
-
-//             setRecipeDetails(mergedData);
-//           } else {
-//             console.log("User not found");
-//           }
-//         } else {
-//           console.log("Recipe not found");
-//         }
-//       } catch (error) {
-//         console.error("Error fetching recipe details:", error.message);
-//       }
-//     };
-
-//     fetchRecipeDetails();
-//   }, [recipeId]);
-
-//   const handleUpdateRecipeClick = () => {
-//     // Redirect the user to the update page with the recipeId
-//     console.log("forwording");
-//     router.push(`/recipes/${recipeId}/update`);
-//   };
-
-//   return (
-//     <div className="container mt-5">
-//       <div className="row">
-//         <div
-//           className="col-md-3"
-//           style={{
-//             padding: "20px",
-//             margin: "10px",
-//             border: "1px solid #ccc",
-//             borderRadius: "5px",
-//           }}
-//         >
-//           {/* Recipe Name, Posted by, User Image */}
-//           <div className="mb-4">
-//             <h2 style={{ fontSize: "24px", fontWeight: "bold" }}>
-//               {recipeDetails?.RecipeName}
-//             </h2>
-//             <p style={{ fontSize: "16px" }}>
-//               Posted by: {recipeDetails?.postedByUserName}
-//             </p>
-//             {/* <img
-//               src={recipeDetails?.userImageUrl}
-//               alt="User"
-//               className="img-fluid rounded-circle"
-//               style={{ width: "50px", height: "50px" }}
-//             /> */}
-//           </div>
-
-//           {/* Recipe Image */}
-//           <div className="mb-3">
-//             <img
-//               src={recipeDetails?.imgUrl}
-//               alt="Recipe"
-//               className="img-fluid"
-//               style={{ width: "300px", height: "200px" }}
-//             />
-//           </div>
-//         </div>
-
-//         <div
-//           className="col-md-6"
-//           style={{
-//             padding: "20px",
-//             margin: "10px",
-//             border: "1px solid #ccc",
-//             borderRadius: "5px",
-//           }}
-//         >
-//           {/* Recipe Ingredients */}
-//           <div className="mb-4">
-//             <h3 style={{ fontSize: "20px", fontWeight: "bold" }}>
-//               Ingredients:
-//             </h3>
-//             <div
-//               style={{
-//                 height: "100px",
-//                 width: "100%",
-//                 fontSize: "16px",
-//                 textAlign: "justify",
-//               }}
-//             >
-//               {recipeDetails?.Ingredients}
-//             </div>
-//           </div>
-
-//           {/* Cooking Instructions */}
-//           <div>
-//             <h3 style={{ fontSize: "20px", fontWeight: "bold" }}>
-//               Cooking Instructions:
-//             </h3>
-//             <p style={{ fontSize: "16px", textAlign: "justify" }}>
-//               {recipeDetails?.Instructions}
-//             </p>
-//           </div>
-
-//           <div className="d-flex mt-3">
-//             <button
-//               className="btn btn-primary flex-grow-1 me-2"
-//               type="button"
-//               onClick={handleUpdateRecipeClick}
-//               style={{ fontSize: "14px" }} // Adjust the font size
-//             >
-//               Update Recipe
-//             </button>
-//             <button
-//               className="btn btn-primary flex-grow-1"
-//               type="button"
-//               style={{ fontSize: "14px" }} // Adjust the font size
-//             >
-//               Delete Recipe
-//             </button>
-//           </div>
-//         </div>
-//       </div>
-//     </div>
-//   );
-// }
-
 
 "use client";
 import { useEffect, useState } from "react";
@@ -173,24 +6,25 @@ import { db } from "../../../firebase";
 import { useRouter } from "next/navigation";
 import { auth } from "../../../firebase";
 import { getAuth } from "firebase/auth";
+import { useAuthState } from "react-firebase-hooks/auth";
 
 export default function RecipeDetails({ params }) {
   const [recipeDetails, setRecipeDetails] = useState(null);
   const recipeId = params.id;
   const router = useRouter();
-  const [currentUserId, setCurrentUserId] = useState(null); // Assuming you have the current user's ID
+  const [currentUserId, setCurrentUserId] = useState(null);
+  const [user, loading, error] = useAuthState(auth);
 
   const getCurrentUserId = () => {
     const auth = getAuth();
     const user = auth.currentUser;
-  
+
     if (user) {
       return user.uid;
     } else {
       return null;
     }
   };
-  
 
   useEffect(() => {
     const fetchRecipeDetails = async () => {
@@ -201,14 +35,12 @@ export default function RecipeDetails({ params }) {
         if (recipeDoc.exists()) {
           const recipeData = recipeDoc.data();
 
-          // Retrieve user information based on userId
           const userDocRef = doc(db, "Users", recipeData.userId);
           const userDoc = await getDoc(userDocRef);
 
           if (userDoc.exists()) {
             const userData = userDoc.data();
 
-            // Merge recipe and user data
             const mergedData = {
               ...recipeData,
               postedByUserName: userData.name,
@@ -216,8 +48,7 @@ export default function RecipeDetails({ params }) {
 
             setRecipeDetails(mergedData);
 
-            // Assuming you have the current user's ID available
-            const userId = getCurrentUserId(); // Replace this with your actual method to get the current user's ID
+            const userId = getCurrentUserId();
             setCurrentUserId(userId);
           } else {
             console.log("User not found");
@@ -234,19 +65,22 @@ export default function RecipeDetails({ params }) {
   }, [recipeId]);
 
   const handleUpdateRecipeClick = () => {
-    // Redirect the user to the update page with the recipeId
-    console.log("forwarding");
-    router.push(`/recipes/${recipeId}/update`);
+    if (user) {
+      router.push(`/recipes/${recipeId}/update`);
+    } else {
+      // Handle the case where the user is not logged in
+      // You can show a pop-up or redirect to the login page
+      alert("Please log in to update the recipe.");
+    }
   };
 
   const handleDeleteRecipeClick = async () => {
-    // Display a confirmation pop-up
     const isConfirmed = window.confirm("Are you sure you want to delete this recipe?");
 
     if (isConfirmed) {
       try {
         // Check if the current user is the one who created the recipe
-        if (currentUserId === recipeDetails.userId) {
+        if (user && currentUserId === recipeDetails.userId) {
           // Delete the recipe (replace the following line with your delete logic)
           await deleteDoc(doc(db, "Recipes", recipeId));
 
@@ -275,7 +109,6 @@ export default function RecipeDetails({ params }) {
             borderRadius: "5px",
           }}
         >
-          {/* Recipe Name, Posted by, User Image */}
           <div className="mb-4">
             <h2 style={{ fontSize: "24px", fontWeight: "bold" }}>
               {recipeDetails?.RecipeName}
@@ -285,7 +118,6 @@ export default function RecipeDetails({ params }) {
             </p>
           </div>
 
-          {/* Recipe Image */}
           <div className="mb-3">
             <img
               src={recipeDetails?.imgUrl}
@@ -305,7 +137,6 @@ export default function RecipeDetails({ params }) {
             borderRadius: "5px",
           }}
         >
-          {/* Recipe Ingredients */}
           <div className="mb-4">
             <h3 style={{ fontSize: "20px", fontWeight: "bold" }}>
               Ingredients:
@@ -322,7 +153,6 @@ export default function RecipeDetails({ params }) {
             </div>
           </div>
 
-          {/* Cooking Instructions */}
           <div>
             <h3 style={{ fontSize: "20px", fontWeight: "bold" }}>
               Cooking Instructions:
@@ -333,13 +163,13 @@ export default function RecipeDetails({ params }) {
           </div>
 
           <div className="d-flex mt-3">
-            {currentUserId === recipeDetails?.userId && (
+            {currentUserId === recipeDetails?.userId && currentUserId && (
               <>
                 <button
                   className="btn btn-primary flex-grow-1 me-2"
                   type="button"
                   onClick={handleUpdateRecipeClick}
-                  style={{ fontSize: "14px" }} // Adjust the font size
+                  style={{ fontSize: "14px" }}
                 >
                   Update Recipe
                 </button>
@@ -347,7 +177,7 @@ export default function RecipeDetails({ params }) {
                   className="btn btn-primary flex-grow-1"
                   type="button"
                   onClick={handleDeleteRecipeClick}
-                  style={{ fontSize: "14px" }} // Adjust the font size
+                  style={{ fontSize: "14px" }}
                 >
                   Delete Recipe
                 </button>
